slight rearranging of instantiation code
