+++ source/rust_verify/src/verifier.rs
@@ -2,6 +2,7 @@ use crate::commands::{Op, OpGenerator, OpKind, QueryOp, Style};
 use crate::config::{Args, ShowTriggers};
 use crate::context::{ArchContextX, ContextX, ErasureInfo};
 use crate::debugger::Debugger;
+use crate::profiler::write_instantiation_graph;
 use crate::spans::{SpanContext, SpanContextX};
 use crate::user_filter::UserFilter;
 use crate::util::error;
@@ -1982,78 +1983,6 @@ fn delete_dir_if_exists_and_is_dir(dir: &std::path::PathBuf) -> Result<(), VirEr
             return Err(error(format!("{} exists and is not a directory", dir.display())));
         }
     })
-}
-
-fn write_instantiation_graph(
-    bucket_id: &BucketId,
-    op: Option<&Op>,
-    func_map: &HashMap<Fun, vir::ast::Function>,
-    profiler: &Profiler,
-    qid_map: &HashMap<String, vir::sst::BndInfo>,
-    profile_file_name: std::path::PathBuf,
-) {
-    let air::profiler::InstantiationGraph { edges, nodes, names } = profiler.instantiation_graph();
-    use tool_facade::*;
-    let name_strs: HashSet<String> = names.values().cloned().collect();
-    let quantifiers: HashMap<String, Quantifier> = name_strs
-        .iter()
-        .map(|n| {
-            let bnd_info = qid_map.get(n);
-            let kind = if n.starts_with(air::profiler::USER_QUANT_PREFIX) {
-                QuantifierKind::User(UserQuantifier {
-                    span: bnd_info.as_ref().unwrap().user.as_ref().unwrap().span.as_string.clone(),
-                })
-            } else {
-                QuantifierKind::Internal
-            };
-            (
-                n.clone(),
-                std::rc::Rc::new(QuantifierX {
-                    qid: n.clone(),
-                    module: bnd_info.map(|b| {
-                        module_name(
-                            &func_map[&b.fun].x.owning_module.as_ref().expect("owning module"),
-                        )
-                    }),
-                    kind,
-                }),
-            )
-        })
-        .collect();
-    let instantiations: HashMap<(u64, usize), Instantiation> = nodes
-        .iter()
-        .map(|n| {
-            (
-                n.clone(),
-                std::rc::Rc::new(InstantiationX {
-                    quantifier: quantifiers[&names[n]].clone(),
-                    id: *n,
-                }),
-            )
-        })
-        .collect();
-    let mut graph: HashMap<Instantiation, HashSet<((), Instantiation)>> = HashMap::new();
-    for (src, tgts) in edges {
-        let graph_src = graph.entry(instantiations[src].clone()).or_insert(HashSet::new());
-        for tgt in tgts {
-            graph_src.insert(((), instantiations[tgt].clone()));
-        }
-    }
-    let instantiations: HashSet<Instantiation> = instantiations.values().cloned().collect();
-    let quantifiers = quantifiers.into_values().collect();
-    let instantiation_graph = InstantiationGraph {
-        bucket_name: bucket_id.to_log_string(),
-        module: module_name(bucket_id.module()),
-        function: op.map(|op| fun_as_friendly_rust_name(&op.get_function().x.name)),
-        quantifiers,
-        instantiations,
-        graph: Graph(graph),
-    };
-    let file_name = profile_file_name.with_extension("graph");
-    let mut f = File::create(&file_name)
-        .expect(&format!("failed to open instantiation graph file {}", file_name.display()));
-    bincode::serialize_into(&mut f, &instantiation_graph)
-        .expect("failed to write instantiation graph");
 }
 
 // TODO: move the callbacks into a different file, like driver.rs
